chat-api refactoring
